create life.go to separate life implementation into another file.
